+++ src/server/gameserver/LoginServerManager.cpp
@@ -94,7 +94,7 @@ void LoginServerManager::run ()
 			port = g_pConfig->getPropertyInt("DB_PORT");
 
 		Connection* pConnection = new Connection(host, db, user, password, port);
-		g_pDatabaseManager->addConnection((int)Thread::self(), pConnection);
+		g_pDatabaseManager->addConnection((int)(long)Thread::self(), pConnection);
 		cout << "************************************************************************" << endl;
 		cout << "************************************************************************" << endl;
 		cout << "************************************************************************" << endl;
